backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
PyJWT
pydantic
pydantic[email]
prometheus-fastapi-instrumentator